requirements.txt
Django==2.0.6
requests==2.19.1
psycopg2==2.7.5
channels==2.1.2
names==0.3.0
djangorestframework==3.8.2
django-filter==2.0.0
Markdown==2.6.11
git+git://github.com/Romeno/chronotrack@master#chronotrack
django-grappelli==2.11.1
django-filebrowser==3.10.1
Pillow==5.2.0
googlemaps==3.0.2
django-flatblocks=0.9.4
django-tinymce=2.7.0
